Update for OpenTracing 0.31.0 API specification
Consolidate SpanImpl into BaseSpanImpl.  Will rename in a subsequent
commit in order to preserve history.
