--- src/main/java/com/ebay/opentracing/basic/SpanImpl.java
@@ -1,43 +1,0 @@
-/*
- * Copyright (c) 2017 eBay Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ebay.opentracing.basic;
-
-import io.opentracing.Span;
-
-import java.util.concurrent.TimeUnit;
-
-final class SpanImpl<T> extends BaseSpanImpl<Span, T> implements Span {
-    private final MutableSpanData<T> spanState;
-    private final SpanFinisher<T> spanFinisher;
-
-    SpanImpl(MutableSpanData<T> spanState, SpanFinisher<T> spanFinisher) {
-        super(spanState);
-        this.spanState = spanState;
-        this.spanFinisher = spanFinisher;
-    }
-
-    @Override
-    public void finish() {
-        spanFinisher.finish(spanState);
-    }
-
-    @Override
-    public void finish(long finishMicros) {
-        spanFinisher.finish(spanState, TimeUnit.MICROSECONDS, finishMicros);
-    }
-
-}
